requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
click==8.1.8
    # via flask
flask==3.1.0
    # via -r requirements.in
importlib-metadata==8.6.1
    # via flask
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mysql-connector-python==9.2.0
    # via -r requirements.in
werkzeug==3.1.3
    # via flask
zipp==3.21.0
    # via importlib-metadata
